requirements.txt
appdirs==1.4.3
backports.csv==1.0.7
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-import-export==1.2.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
gunicorn==20.0.0
idna==2.8
jdcal==1.4.1
lxml==4.3.4
odfpy==1.4.0
openpyxl==2.6.2
parse==1.12.0
psycopg2==2.8.2
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
tqdm==4.32.2
urllib3==1.25.3
w3lib==1.20.0
websockets==8.0
xlrd==1.2.0
xlwt==1.3.0
